separate out auto tests
